New Files and HTML Basics
Made the files that will be necessary for the capstone project, and added their HTML skeletons. Also added a reset file for the project that set all the styles for the basic HTML elements. Added a basic navigation bar that has the links to the other pages, which do not yet have code inside them.