Deleted heatherindex Branch, Structure Layout
Merged the heatherindex file into the master branch where Eli and I are now both working. Changed the container for the footer, and justified the content to be in the center. Changed the styling for the updates and the new gear sections of the index. Made almost all of the text on the index page centered.