Mission Statement Spacing, Darken Filter Nav Links
Wrote a transition in CSS that allowed the nav bar links to darken slowly when they were hovered over (but not using the hover feature, in order to make the animation smoother). Also changed the CSS padding that changes the way the mission statement is seen, along with the height so that the page is filled by this.