Rain Icon, Footer Changes
I changed the rain icon to simplofy it because Eli hates my desgins, and also changed the footer social media links because Eli hates that all the items on the page change in color when you hover over them. Now they're just grey because they look trashy otherwise. And the other footer change was the height because for some damn reason it will just never ever ever ever no matter what I do stay the height and look the way that it is supposed to! Wow! Also I deleted that cursed birb picture because like, we should'nt have that thing anymore, at all.