Fixed Footer
I think I fixed that stupid width for the footer, which never works! It's fine locally but may not work hosted, oh well. Also there were merge conflicts that needed to be resolved because the merges were seen in the code with that arrow thingy, so I just took that out. Entirely possible I deleted someones code. Meh.