Gradients and Images
Changed the colors of most of the images that we used so that we can't be sued for copyright, also added logos to each product so we Might actually be sued. Also fixed the issue with the gradient and the footer by decreasing the opacity of the footer and the gradient end color.