Form Changes, Contact Content
Added content for each one of the contact people thingies for the contact page. Went to W3C and found the most basic of all forms to exist and implemented (not copied exactly) the form into our code. Used inline CSS to get the size of the biggect text box, put placeholders in each, and put the boxes centered in the middle of the page. Also changed the colors of the links on the contact page both in the nav bar (white to black) and in the emails (now hover blue to white).